Add support for alternate download method
This patch adds support for injecting other download methods than clone
to be used to download dependencies. It will allow to use git submodule
add instead of git clone for example.

It also reduces the amount of monkeypatching needed to test the download
operation. This is good since monkeypatching is manipulating global
state and this should be avoided (patching decorators and context
managers mitigate this issue).