Skip dependencies with no package.yml
This patch adds the ability to simply skip dependencies for which no
package.yml file can be found.

This was needed because Travis CI will need to be able to build
applications depending on uc/OS-III but we cannot give public access to
this repository since it is a closed source project.
Packager will simply skip uc/OS-III package now, since it is not needed
to build and run unit tests.