Fix CFLAGS append
It wasn't possible to pass CFLAGS from the environment before.
For example running CFLAGS="-Werror" resulted in the compiler invocation
looking like this: "gcc -Werror; -Wall foo.c" which broke the build.
From the CMake mailing list:

SET(VAR a b) leads to: a;b
SET var "a b") results in: a b