Allow disabling generation of unittest CMakeLists
Some boards now have a manually written CMakeLists.txt as we try to
migrate away from package.yml-based builds. This means that in those
cases, running packager overwrites a manually written file, possibly
erasing uncommitted changes.

To avoid this, we add a setting in the package.yml that can be used to
disable the default behavior of generating the CMakeLists.txt if the
project contains any unit tests (which probably was a bad feature to
begin with).